requirements.txt
discord.py
pytz
python-decouple
aiohttp
